feat(compaction): move iceberg_compaction_event_loop to manager
