+++ src/meta/node/src/server.rs
@@ -368,9 +368,6 @@ pub async fn start_service_as_election_leader(
     let (compactor_streams_change_tx, compactor_streams_change_rx) =
         tokio::sync::mpsc::unbounded_channel();
 
-    let (iceberg_compactor_streams_change_tx, iceberg_compactor_streams_change_rx) =
-        tokio::sync::mpsc::unbounded_channel();
-
     let meta_metrics = Arc::new(GLOBAL_META_METRICS.clone());
 
     let hummock_manager = hummock::HummockManager::new(
@@ -379,7 +376,6 @@ pub async fn start_service_as_election_leader(
         meta_metrics.clone(),
         compactor_manager.clone(),
         compactor_streams_change_tx,
-        iceberg_compactor_streams_change_tx,
     )
     .await
     .unwrap();
@@ -482,9 +478,10 @@ pub async fn start_service_as_election_leader(
     let iceberg_compactor_manager = Arc::new(IcebergCompactorManager::new());
 
     // TODO: introduce compactor event stream handler to handle iceberg compaction events.
-    let (iceberg_compaction_mgr, _iceberg_compactor_event_rx) = IcebergCompactionManager::build(
+    let (iceberg_compaction_mgr, iceberg_compactor_event_rx) = IcebergCompactionManager::build(
         metadata_manager.clone(),
         iceberg_compactor_manager.clone(),
+        meta_metrics.clone(),
     );
 
     sub_tasks.push(IcebergCompactionManager::compaction_stat_loop(
@@ -619,9 +616,9 @@ pub async fn start_service_as_election_leader(
         compactor_streams_change_rx,
     ));
 
-    sub_tasks.extend(HummockManager::iceberg_compaction_event_loop(
-        hummock_manager,
-        iceberg_compactor_streams_change_rx,
+    sub_tasks.extend(IcebergCompactionManager::iceberg_compaction_event_loop(
+        iceberg_compaction_mgr.clone(),
+        iceberg_compactor_event_rx,
     ));
 
     sub_tasks.push(
